# storage/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sistema Académico API",
    "description": "API REST del Sistema de Gestión Académica. Endpoints para autenticación, administración, gestión de usuarios, materias, grados, secciones, horarios, módulos, tareas, parciales, entregas y más.\n\n## Autenticación\n\nLa API utiliza Laravel Sanctum para autenticación. Para acceder a endpoints protegidos:\n\n1. Envíe credenciales al endpoint POST /api/login\n2. Reciba un token JWT en la respuesta\n3. Incluya el token en el header: Authorization: Bearer <token>\n\n## Roles\n\n- Admin: Acceso completo a todas las funcionalidades\n- Profesor: Gestión de materias, módulos, tareas, parciales y calificaciones\n- Estudiante: Visualización de materias, tareas y entregas\n\n## Códigos de Respuesta\n\n| Código | Descripción |\n|--------|-------------|\n| 200 | Éxito |\n| 201 | Creado exitosamente |\n| 400 | Solicitud incorrecta |\n| 401 | No autenticado |\n| 403 | Sin permisos |\n| 404 | No encontrado |\n| 409 | Conflicto |\n| 422 | Error de validación |\n| 429 | Demasiadas solicitudes |\n| 500 | Error del servidor",
    "version": "1.0.0",
    "contact": {
      "name": "Desarrollo",
      "email": "admin@academico.local"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {"type": "string", "format": "email", "example": "admin@admin.com"},
          "password": {"type": "string", "example": "admin123"}
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {"type": "string", "example": "eyJ0eXAiOiJKV1QiLCJhbGc..."},
          "token_type": {"type": "string", "example": "Bearer"},
          "user": {"type": "object"}
        }
      },
...
    }
  },
  "paths": {
    "/api/login": {
      "post": {
        "tags": ["Autenticacion"],
        "summary": "Iniciar sesion",
        "description": "Autentica un usuario y retorna un token de acceso",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/LoginRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/LoginResponse"}
              }
            }
          },
          "401": {"description": "Credenciales inválidas"},
          "422": {"description": "Error de validación"},
          "429": {"description": "Demasiados intentos"}
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": ["Autenticacion"],
        "summary": "Cerrar sesión",
        "description": "Invalida el token de acceso actual",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Logout exitoso"},
          "401": {"description": "No autenticado"}
        }
      }
    },
...
  }
}